requirements_test.txt
debugpy        
pytest         
pytest-cov     
python-dotenv  
psycopg2-binary
requests
